contract-intel/backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
python-multipart>=0.0.9
anthropic>=0.28.0
pypdf>=4.0.0
python-dotenv>=1.0.0
